\name{coeftab}
\alias{coeftab}
\title{Coefficient tables}
\description{
  Returns a table of model coefficients in rows and models in columns.
}
\usage{
coeftab( ... , intercept=TRUE )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{A series of inla fit models, separated by commas}
  \item{intercept}{If TRUE, intercepts are included.}
}
\details{
  This function provides a way to compare estimates across models.
}
\value{
}
\references{}
\author{Dong Liang}
\seealso{}
\examples{
}
\keyword{ }

