% Generated by roxygen2: do not edit by hand
\name{DIC}
\alias{DIC}
\title{Deviance information criterion
 
Estimate deviance information criterion from fitted INLA objects}
\usage{
DIC(...)
}
\arguments{
\item{...}{A list of fitted INLA objects}
}
\value{
A data frame with the DIC from each model}
\description{
Model comparison using Deviance Information Criterion.
}
\author{
Dong Liang <dliang@umces.edu>
}
