\name{PI}
\alias{PI}

\title{Confidence/credible intervals from samples}
\description{
  These functions compute highest posterior density (HPDI) and percentile (PI) intervals, using samples from a posterior density or simulated outcomes..
}
\usage{
PI (samples, prob = 0.89) 
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{samples}{Vector of parameter values}
  \item{prob}{interval probability mass}
}
\details{
  TBD
}
\value{
}
\references{
}
\author{Dong Liang}
\examples{
\dontrun{
}
}
\keyword{}

