% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm_inla.R
\name{sgam}
\alias{sgam}
\title{generalized linear model for spatial data}
\usage{
sgam(formula,data,newdata=NULL,param=NULL,...)
}
\arguments{
\item{formula}{lme type formula object with additive terms}

\item{data}{data frame to evaluate the model}

\item{newdata}{data frame to form prediction}

\item{param}{Hyper-parameters for multivariate normal prior for random coefficient}

\item{...}{Additional parameter passed to the \code{inla} algorithm}
}
\value{
fitted \code{inla} object
}
\details{
GAMM using R and Integrated Nested Laplace Approximation.
}
\author{Dong Liang}
