% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm_inla.R
\name{sglm}
\alias{sglm}
\title{generalized linear model for spatial data}
\usage{
sglm(formula, data, newdata = NULL, family = "gaussian",
  E = NULL, Ntrials = NULL, param = NULL, weights = NULL,
  verbose = FALSE)
}
\arguments{
\item{formula}{lme type formula object with additive terms}

\item{data}{data frame to evaluate the model}

\item{newdata}{data frame to form prediction}

\item{family}{family for response variable}

\item{E}{offset term}

\item{Ntrials}{number of trials INLA term}

\item{param}{Hyper-parameters for multivariate normal prior for random coefficient}

\item{weights}{ weights in weighted least squars.}

\item{verbose}{whether to show detailed results}
}
\value{
fitted foo object
}
\description{
GAMM using R and Integrated Nested Laplace Approximation.
}
