\name{sp713-interval}
\alias{sp713-interval}
\alias{col.alpha}
\alias{set_nice_margins}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal functions to sp713
}
\description{
These functions are working R functions that are called by the sglm function. They should not be directly used.
}

