% Generated by roxygen2: do not edit by hand
\name{visio}
\alias{visio}
\title{Leaflet mapping of spatial data}
\usage{
visio(
  obj,leg.col="white",field=NULL,
  choropleth=TRUE,bins=NA,nquantile.color=4,na.color="#808080",
  pt.cex=4,pt.col="black",pt.weight=1,
  project=TRUE,Graticule=FALSE,
  legend=TRUE,legend.position="bottomright",legend.title="",
  add=NULL,
  ...)
}
\arguments{
\item{obj}{a simple feature object to visualize}
\item{leg.col}{color palette name from a Color Brewer}
\item{field}{the field name to visualize, can be numeric or categorical}
\item{choropleth}{logical, whether showing choropleth or numeric color}
\item{bins}{for choropleth a vector of color dividers}
\item{nquantile.color}{for choropleth default number of quantiles to divide the data for color definitions}
\item{na.color}{The color to return for NA values}
\item{pt.cex}{for point data the size of the point}
\item{pt.col}{for point data the color of the outline}
\item{pt.weight}{for point data the weight/width of the outline}
\item{project}{whether to project raster to EPSG:3857}
\item{Graticule}{whether to add graticule}
\item{legend}{whether to include legend}
\item{legend.title}{title of the legend}
\item{legend.position}{default legend posistion}
\item{add}{add to an existing layer, not used}
\item{...}{additional arguments to graphic layer code}
}
\value{
A leaflet map}
\details{
The simple feature should be in WGS 84 projection}
\author{
Dong Liang <dliang@umces.edu>
}
